extracted `ComputeStrokes` into its own class
